+++ docs/RELEASENOTES.md
@@ -11,7 +11,7 @@ Release type: **Major** / Minor / Hotfix
 - API
   * void afStart(String key, boolean isDebug)
   * void afSetCustomerUserId(String id)
-  * void void afTrackEvent(String eventName, Map<String, Object> eventValues)
+  * void void afLogEvent(String eventName, Map<String, Object> eventValues)
   
 - Android SDK to v4.9.0